created settings layout
